change rank int to long

+++ src/IMS/core/IMS.BusinessModel/Dto/BaseDto.cs
@@ -14,8 +14,8 @@ namespace IMS.BusinessModel.Dto
         public virtual DateTime CreationDate { get; set; }
         public virtual T ModifyBy { get; set; }
         public virtual DateTime? ModificationDate { get; set; }
-        public virtual int Rank { get; set; }
-        public virtual int? VersionNumber { get; set; }
+        public virtual long Rank { get; set; }
+        public virtual long? VersionNumber { get; set; }
         public virtual string BusinessId { get; set; }
     }
 }
